+++ src/platform.test.ts
@@ -34,7 +34,7 @@ describe('TestPlatform', () => {
       matterbridgeDirectory: './jest/matterbridge',
       matterbridgePluginDirectory: './jest/plugins',
       systemInformation: { ipv4Address: undefined },
-      matterbridgeVersion: '2.2.5',
+      matterbridgeVersion: '3.0.3',
       getDevices: jest.fn(() => {
         // console.log('getDevices called');
         return [];
@@ -71,7 +71,6 @@ describe('TestPlatform', () => {
       'featureBlackList': ['device_temperature', 'update', 'update_available', 'power_outage_count', 'indicator_mode', 'do_not_disturb', 'color_temp_startup'],
       'deviceFeatureBlackList': {},
       'postfix': '',
-      'postfixHostname': true,
       'debug': true,
       'unregisterOnShutdown': false,
     } as PlatformConfig;